requirements.txt
django>=3.2.21,<3.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.7,<3
drf-spectacular>=0.26.5,<0.27